requirements.txt
# --- Framework & Server ---
flask==3.1.2
waitress==3.0.2

# --- Network & API ---
requests==2.32.5

# --- Tokenization (Logic) ---
tiktoken==0.12.0
mistral_common==1.8.5

# --- File Type Detection ---
filetype==1.2.0

# --- Security & Rate Limiting ---
flask-limiter==4.1.1
limits[memcached]==5.6.0
pymemcache==4.0.0

# --- Logging & Aesthetics ---
coloredlogs==15.0.1
printedcolors==1.0.0

# --- Environment & Configuration ---
python-dotenv==1.2.1

# --- Testing ---
pytest==9.0.2
pytest-cov>=4.1.0

# --- Development Tools ---
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
bandit>=1.7.5
pre-commit>=3.0.0
detect-secrets>=1.4.0
